Add Cluster Type dropdown field in Project description

+++ marlo-web/src/main/resources/database/migrations/V2_6_0_20210601_1100__CreateTableClusterTypes.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `cluster_types` (
-`id`  int(20) NOT NULL AUTO_INCREMENT ,
+`id`  bigint(20) NOT NULL AUTO_INCREMENT ,
 `name`  text NULL ,
 PRIMARY KEY (`id`),
 INDEX `id` (`id`) USING BTREE
